Extract e2e testing facilities in separate package
