+++ tests/end_to_end/bpftest/__init__.py
@@ -1,0 +1,8 @@
+# SPDX-License-Identifier: GPL-2.0
+# Copyright (c) 2023 Meta Platforms, Inc. and affiliates.
+
+from .main import BpfilterEndToEndTest
+
+__all__ = [
+    "BpfilterEndToEndTest",
+]
